5_RAG_Chatbot/requirements.txt
streamlit
openai
psycopg2-binary
redis
pdfplumber
pyyaml